feat: add filters.ts helper
